KNN Implementaion
In this implementation, the K-Nearest Neighbors (KNN) algorithm is applied using the Iris dataset from the scikit-learn library. The process begins by identifying its features using the .feature_names attribute. The data and target values are then extracted and stored in variables . Next, the KNeighborsClassifier is imported and initialized. Initially, the entire dataset is fitted to the classifier. To evaluate the model effectively, the dataset is split into training and testing sets based on a specified test size. The training data is used to train the classifier, after which its performance is assessed. The evaluation includes calculating the accuracy score to provide a measure of the prediction's accuracy and generating the confusion matrix to analyze the classification output.