Best Model Analysis
In this implementation, the Wine dataset from the scikit-learn library is used to compare the performance of various machine learning models. The dataset is first loaded and split into features and target variables. A 10-fold cross-validation strategy is then configured using the KFold function, where the dataset is divided into 10 parts . Six models, the Logistic Regression, Decision Tree, K-Nearest Neighbors (KNN), Linear Discriminant Analysis (LDA), Naive Bayes and Support Vector Machine (SVM), are to be compared. Each model is evaluated using cross-validation and the average accuracy along with the standard deviation for each model is obtained to assess performance. The model with the highest average accuracy is  the best performing model. Then a boxplot is generated to visualize the accuracy scores of all the models. Generally the best model depends on the dataset size, feature distribution, feature scaling, class distribution and noises in datasets. LDA is observed to be the best performing model for the applied dataset in this case.